Add:update archive access page for split list-result to components
